\encoding{UTF-8}
\name{month.stats.int}
\title{Internal function for month.stats}
\description{Internal function for \code{\link{month.stats}}}
\arguments{
  \item{data}{Data vector}
  \item{fun}{Function}
  \item{ts}{Time stamp}
  \item{years}{Vector of measurement period}
  \item{digits}{Number of decimal places.}
}
\value{Returns data frame with monthly statistics.}
\author{Christian Graul}
\seealso{\code{\link{month.stats}}}
\keyword{methods}
\keyword{internal}
