\encoding{UTF-8}
\name{monthStatsInt}
\alias{monthStatsInt}
\title{Internal function for month.stats}
\description{Internal function for \code{\link{monthStats}}}
\usage{monthStatsInt(data, ts, years)}
\arguments{
  \item{data}{Data vector}
  \item{ts}{Time stamp}
  \item{years}{Vector of measurement period}
}
\value{Returns data frame with monthly means.}
\author{Christian Graul}
\seealso{\code{\link{monthStats}}}
\keyword{methods}
\keyword{internal}
