\encoding{UTF-8}
\name{subset.int}
\alias{subset.int}
\title{Internal function data subsets}
\description{Internal function for data subsets}
\usage{subset.int(timestamp, subset)}
\arguments{
  \item{timestamp}{Time stamp}
  \item{subset}{Subset, containing start and end time stamp.}
}
\value{Returns index of start and end time stamp.}
\author{Christian Graul}
\seealso{\code{\link{aep}}, \code{\link{availability}}, \code{\link{energy}}, \code{\link{frequency}}, \code{\link{month.stats}}, \code{\link{day.plot}}, \code{\link{polar.plot}}, \code{\link{turb.iec.plot}}, \code{\link{profile}}, \code{\link{turbulence}}, \code{\link{weibull}}}
\keyword{methods}
\keyword{internal}
