\encoding{UTF-8}
\name{subsetInt}
\alias{subsetInt}
\title{Internal function data subsets}
\description{Internal function for data subsets}
\usage{subsetInt(time.stamp, subset)}
\arguments{
  \item{time.stamp}{Time stamp}
  \item{subset}{Subset, containing start and end time stamp.}
}
\value{Returns index of start and end time stamp.}
\author{Christian Graul}
\seealso{\code{\link{aep}}, \code{\link{availability}}, \code{\link{energy}}, \code{\link{frequency}}, \code{\link{monthStats}}, \code{\link{plotDay}}, \code{\link{plotPolar}}, \code{\link{plotProfile}}, \code{\link{plotTimeSeries}}, \code{\link{plotTurbIEC}}, \code{\link{plotWeibull}}, \code{\link{profile}}, \code{\link{turbulence}}, \code{\link{weibull}}}
\keyword{methods}
\keyword{internal}
