\encoding{UTF-8}
\name{weibullInt}
\alias{weibullInt}
\title{Internal function for weibull}
\description{Internal function for \code{\link{weibull}}}
\usage{weibullInt(v.dat, msg)}
\arguments{
  \item{v.dat}{Vector of wind speed data.}
  \item{msg}{If \code{TRUE}, messages are printed.}
}
\value{Returns the shape parameter k and scale parameter A of the weibull distribution.}
\author{Christian Graul}
\seealso{\code{\link{weibull}}}
\keyword{methods}
\keyword{internal}
