\encoding{UTF-8}
\name{monthStatsInt}
\alias{monthStatsInt}
\title{Internal function for month.stats}
\description{Internal function for \code{\link{monthStats}}}
\usage{monthStatsInt(data, fun, ts, years, digits)}
\arguments{
  \item{data}{Data vector}
  \item{fun}{Function}
  \item{ts}{Time stamp}
  \item{years}{Vector of measurement period}
  \item{digits}{Number of decimal places.}
}
\value{Returns data frame with monthly statistics.}
\author{Christian Graul}
\seealso{\code{\link{monthStats}}}
\keyword{methods}
\keyword{internal}
